chore: extract proposal creation from routes to ProposalCreator
